feat: connect service/repository to api controller(for user)

+++ src/main/java/com/cheer/cheerchat/service/UserService.java
@@ -6,11 +6,10 @@ import com.cheer.cheerchat.request.UpdateUserRequest;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-
 @Service
 public interface UserService {
     public User findUserById(Integer id) throws UserException;
     public User findUserByProfile(String jwt) throws UserException;
-    public User updateUser(Long userid, UpdateUserRequest request) throws UserException;
+    public User updateUser(Integer userid, UpdateUserRequest request) throws UserException;
     public List<User> searchUser(String query);
 }
